feat: improves builders page retrieve by introducing a manifest
