+++ .gitignore
@@ -21,3 +21,6 @@ node_modules
 
 # cli
 dist
+
+# generated files
+packages/nextjs/app/_data/builders.manifest.json
