+++ packages/nextjs/package.json
@@ -3,11 +3,12 @@
   "private": true,
   "version": "0.1.0",
   "scripts": {
-    "build": "next build",
+    "build": "npx tsx scripts/generate-builders-manifest.ts && next build",
     "check-types": "tsc --noEmit --incremental",
-    "dev": "next dev",
+    "dev": "concurrently \"next dev\" \"npx tsx scripts/watch-builders-manifest.ts\"",
     "format": "prettier --write . '!(node_modules|.next|contracts)/**/*'",
     "lint": "next lint",
+    "prebuild": "npx tsx scripts/generate-builders-manifest.ts",
     "serve": "next start",
     "start": "next dev",
     "vercel": "vercel --build-env YARN_ENABLE_IMMUTABLE_INSTALLS=false --build-env ENABLE_EXPERIMENTAL_COREPACK=1 --build-env VERCEL_TELEMETRY_DISABLED=1",
@@ -44,6 +45,8 @@
     "@types/react": "^19.0.7",
     "abitype": "1.0.6",
     "bgipfs": "^0.0.12",
+    "chokidar": "^3.6.0",
+    "concurrently": "^8.2.2",
     "eslint": "^9.23.0",
     "eslint-config-next": "^15.2.3",
     "eslint-config-prettier": "^10.1.1",
@@ -51,6 +54,7 @@
     "postcss": "^8.4.45",
     "prettier": "^3.5.3",
     "tailwindcss": "^4.1.3",
+    "tsx": "^4.7.0",
     "type-fest": "^4.26.1",
     "typescript": "^5.8.2",
     "vercel": "^39.1.3"
